+++ tests/progress.rs
@@ -49,6 +49,7 @@ fn tests() {
     t.pass("tests/pass/44-cow-unsized.rs");
     t.pass("tests/pass/45-borrowed-references.rs");
     t.pass("tests/pass/46-datetime-timezones.rs");
+    t.pass("tests/pass/47-nonzero-numerics.rs");
 
     // These files should fail to compile.
     t.compile_fail("tests/fail/96-fail-derive-on-union.rs");